Change autocomplete dir
