+++ src/AutoComplete/AutoComplete.driver.js
@@ -1,5 +1,1 @@
-import inputWithOptionsDriverFactory from '../InputWithOptions/InputWithOptions.driver';
-
-const autoCompleteDriverFactory = inputWithOptionsDriverFactory;
-
-export default autoCompleteDriverFactory;
+export {autocompleteDriverFactory} from 'wix-ui-backoffice/dist/src/components/Autocomplete/Autocomplete.driver';
