+++ src/AutoComplete/AutoComplete.protractor.driver.js
@@ -1,10 +1,1 @@
-const autoCompleteDriverFactory = component => ({
-  click: () => component.click(),
-  getInput: () => component.$(`input`),
-  getDropdown: () => component.$(`[data-hook="dropdown-layout-options"]`),
-  getDropdownItem: index => component.$$(`[data-hook="dropdown-layout-options"] div`).get(index).getText(),
-  getDropdownItemsCount: () => component.$$(`[data-hook="dropdown-layout-options"] div`).getText().count(),
-  element: () => component
-});
-
-export default autoCompleteDriverFactory;
+export {autocompleteDriverFactory} from 'wix-ui-backoffice/dist/src/components/Autocomplete/Autocomplete.protractor.driver';
