--- src/AutoComplete/README.TESTKIT.md
@@ -1,5 +1,0 @@
-# Autocomplete component
-
-> Input with auto suggestions
-
-Autocomplete testkit is identical to [InputWithOptions](https://wix.github.io/wix-style-react/?selectedKind=Core&selectedStory=InputWithOptions&full=0&down=0&left=1&panelRight=0) test kit
