--- src/StylableAutocomplete/Autocomplete.protractor.driver.js
@@ -1,1 +1,0 @@
-export {autocompleteDriverFactory} from 'wix-ui-backoffice/dist/src/components/Autocomplete/Autocomplete.protractor.driver';
