--- src/StylableAutocomplete/index.js
@@ -1,1 +1,0 @@
-export {Autocomplete} from 'wix-ui-backoffice/Autocomplete';
