+++ stories/AutoComplete/ExampleStandard.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Autocomplete} from 'wix-style-react/StylableAutocomplete';
+import {Autocomplete} from 'wix-style-react/Autocomplete';
 
 const style = {
   display: 'inline-block',
@@ -33,4 +33,4 @@ export default () =>
       Error<AutoComplete error errorMessage="This is an error message" options={rtlOptions}/>
     </div>
   </div>;
-  
+
